add flows
